pip_requirements.txt
paho-mqtt
influxdb
flask
flask_cors
flask_socketio
serial
